Criado a classe DataBaseException e ResourceNotFoundException, e atualizado a classe StoreService com as exceções

+++ src/main/java/com/projeto/discount/services/StoreService.java
@@ -2,7 +2,10 @@ package com.projeto.discount.services;
 
 import com.projeto.discount.entities.Store;
 import com.projeto.discount.repository.StoreRepository;
+import com.projeto.discount.services.exceptions.DataBaseException;
+import com.projeto.discount.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,19 +17,26 @@ public class StoreService {
     @Autowired
     private StoreRepository repository;
 
-    //Retornar todos jogos
+    //Retornar todos jogos!
     public List<Store> findAll(){
         return repository.findAll();
     }
 
-    //Retornar jogo por ID
-    public Optional<Store> findByCodigogame(Long codigoGame){
+    //Retornar jogo por ID!
+    public Store findByCodigogame(Long codigoGame){
         Optional<Store> obj = repository.findById(codigoGame);
-        return obj;
+        return obj.orElseThrow(() -> new ResourceNotFoundException(codigoGame)); // Exceção caso não exista no banco de dados o codigogame informado!
     }
 
-    //Deletar por ID
+    //Deletar por ID!
     public void deleteByCodigoGame(Long codigoGame){
-        repository.deleteById(codigoGame);
+        try{
+            if(!repository.existsById(codigoGame)) throw new ResourceNotFoundException(codigoGame);// Verifica se existe o codigogame informado acima, se existir continua, se não Lança exceção!
+            repository.deleteById(codigoGame);
+        }catch (ResourceNotFoundException e){
+            throw new ResourceNotFoundException(codigoGame);
+        }catch (DataIntegrityViolationException e){// Exceção caso ocorra algum conflito com o banco de dados!
+            throw new DataBaseException(e.getMessage());
+        }
     }
 }
